#914 refactor: extract reusable product presentation
