+++ src/components/search-filters/SearchFilterConfigurationGrid.tsx
@@ -15,8 +15,8 @@ import {
 } from "lucide-react";
 import type { ReactNode } from "react";
 import { Badge } from "@/components/ui/badge.tsx";
-import { StatusBadge } from "@/components/product/badges/StatusBadge.tsx";
-import { ShopTypeBadge } from "@/components/product/badges/ShopTypeBadge.tsx";
+import { StatusBadge } from "@/features/product/catalog/components/badges/StatusBadge.tsx";
+import { ShopTypeBadge } from "@/features/product/catalog/components/badges/ShopTypeBadge.tsx";
 import { SHOP_TYPES } from "@/data/internal/shop/ShopType.ts";
 import { PRODUCT_STATES } from "@/data/internal/product/ProductState.ts";
 import {
