+++ src/components/search-filters/SearchFilterSummary.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
-import { StatusBadge } from "@/components/product/badges/StatusBadge.tsx";
-import { ShopTypeBadge } from "@/components/product/badges/ShopTypeBadge.tsx";
+import { StatusBadge } from "@/features/product/catalog/components/badges/StatusBadge.tsx";
+import { ShopTypeBadge } from "@/features/product/catalog/components/badges/ShopTypeBadge.tsx";
 import { Badge } from "@/components/ui/badge.tsx";
 import {
     type SearchFilterArguments,
