+++ src/components/watchlist/WatchlistResults.tsx
@@ -1,5 +1,5 @@
-import { ProductCard } from "@/components/product/overview/ProductCard.tsx";
-import { ProductCardSkeleton } from "@/components/product/overview/ProductCardSkeleton.tsx";
+import { ProductCard } from "@/features/product/catalog/components/cards/ProductCard.tsx";
+import { ProductCardSkeleton } from "@/features/product/catalog/components/cards/ProductCardSkeleton.tsx";
 import { useInView } from "react-intersection-observer";
 import { useEffect } from "react";
 import { H1 } from "@/components/typography/H1.tsx";
