+++ src/features/product/catalog/components/badges/AuctionWindowBadge.tsx
