+++ src/features/product/catalog/components/badges/__tests__/AuctionWindowBadge.test.tsx
