+++ src/features/product/catalog/components/media/ProhibitedImagePlaceholder.tsx
