+++ src/features/product/catalog/components/media/__tests__/ProductCardImageCarousel.test.tsx
